Move capturer classes to separate files
