+++ Baby MonitorTests/Mocks/VideoCapturerMock.swift
@@ -1,0 +1,11 @@
+//
+//  VideoCapturerMock.swift
+//  Baby MonitorTests
+//
+
+@testable import BabyMonitor
+
+struct VideoCapturerMock: VideoCapturer {
+    func resumeCapturing() {}
+    func stopCapturing() {}
+}
